moved two methods to helper folder
