+++ backend/crypto/ethereum/views.py
@@ -3,44 +3,35 @@ from django.http import JsonResponse, HttpResponse
 import requests
 import json
 import os
-# from dotenv import load_dotenv
+from helper.add_decimals_to_number import *
+from helper.get_symbols_from_dictionary import *
 
-# localhost:8000/ethereum/wallet-balance/0xdB3c617cDd2fBf0bb4309C325F47678e37F096D9
-
-# Create your views here.
-
-# First Param is a number, e.g. 123456789
-# Second Param is a number, relating to how many decimals your number should resolve to, e.g. 5
-# Return Value = 1234.56789
-def create_decimal_from_number(number, decimals):
-    return float(number) / pow(10, int(decimals))
 
 def get_moralis_erc20(address):
+    # What does this method do?
+
+    # It takes in an address from the front end (if "ethereum") is selected
+    # it returns a list of dictionaries, each dictionary represents an ERC token and its balances
+
     url = f"https://deep-index.moralis.io/api/v2/{address}/erc20?chain=eth"
         
     headers = {
-    'x-api-key': os.environ['MORALIS_API_KEY']
+        'x-api-key': os.environ['MORALIS_API_KEY']
     }
 
     response = requests.request("GET", url, headers=headers)        
-    tokens = json.loads(response.text)
+    tokens = json.loads(response.text) #Example Data, tokens: [{'token_address': '0xd2dda223b2617cb616c1580db421e4cfae6a8a85', 'name': 'Bondly Token', 'symbol': 'BONDLY', 'logo': 'https://cdn.moralis.io/eth/0xd2dda223b2617cb616c1580db421e4cfae6a8a85.png', 'thumbnail': 'https://cdn.moralis.io/eth/0xd2dda223b2617cb616c1580db421e4cfae6a8a85_thumb.png', 'decimals': '18', 'balance': '993397116522580432404'}, {'token_address': '0x43901e05f08f48546fff8d6f8df108f60570498b', 'name': 'BASENJI', 'symbol': 'BSJ', 'logo': None, 'thumbnail': None, 'decimals': '18', 'balance': '50000000000000000000'}]
     token_list = []
 
     for token in tokens:
         token_dict = {}
         token_dict['token'] = token['symbol']
-        token_dict['quantity'] = create_decimal_from_number(token['balance'], token['decimals'])
         token_dict['name'] = token['name']
+        token_dict['quantity'] = create_decimal_from_number(token['balance'], token['decimals'])
         token_list.append(token_dict)
 
+    # Example Data, token_list: [{'token': 'BONDLY', 'name': 'Bondly Token', 'quantity': 993.3971165225804}, {'token': 'BSJ', 'name': 'BASENJI', 'quantity': 50.0}]
     return token_list
-
-def get_list_of_symbols(array_of_token_dictionaries):
-    array_of_symbols = []
-    separator = ','
-    for token in array_of_token_dictionaries:
-        array_of_symbols.append(token['token'])
-    return separator.join(array_of_symbols)
 
 def get_cryptocompare_token_price_by_id(token_data):
     url = "https://min-api.cryptocompare.com/data/pricemulti?fsyms=" + get_list_of_symbols(token_data) + "&tsyms=USD"
@@ -51,7 +42,6 @@ def get_cryptocompare_token_price_by_id(token_data):
             token['USDperUnit'] = token_prices[token['token']]['USD']
         except:
             token['USDperUnit'] = "N/A"
-    print(token_data)
     return token_data
 
 def get_token_current_value_in_USD(token_list):
@@ -62,7 +52,7 @@ def get_token_current_value_in_USD(token_list):
             token['BalanceInUSD'] = "N/A"
     return token_list
 
-def get_wallet_balance(request, address):
+def get_ethereum_and_erc20_wallet_balance(request, address):
     tokens = get_moralis_erc20(address) #refactor into next line
     tokens = get_cryptocompare_token_price_by_id(tokens)
     tokens = get_token_current_value_in_USD(tokens)
